Refactor messaging and postback handling logic
